git-svn-id: https://svn.lcsr.jhu.edu/cisst/branches/2012-05-04-matlab@4612 0959eb1a-6956-4f60-947b-6092a4e22c97

+++ cisst/cisstVector/examples/tutorial/ExampleSTL.h
@@ -23,7 +23,7 @@ void ExampleSTL(void) {
     typedef vctFixedSizeVector<double, 6> VectorType;
 
     VectorType vector1;
-    int value = vector1.size();
+    size_t value = vector1.size();
     const VectorType::iterator end = vector1.end();
     VectorType::iterator iter;
     // fill with numbers using iterators
